Change the way we handle internal timer ticks and retrieving timestamps.  The PPS timer now only increments the seconds count.  Timestamp retrieval gets the actual millisecond count from the timer inside a block with interrupts disabled.  Also handles the potential race of timer overflow and disabling interrupts.  This seems to have addressed the occasional glitching in the LED speed delta display which may have been caused by timer values have one less second than expected on occasion.
All internal data values are now stored as floating point 64bit doubles.  The Teensy4 has native 64bit float support, so math performance is fine (beyond
being a 600Mhz+ processor).  Almost all conversions to/from internal data are handled by a set of macros, to make any future adjustments straight
forward.  For output, all numbers are rounded to 3 digits of precision.  Note they are rounded, not truncated.  Timestamps changed from a
{seconds, millis} structure to a double, which makes a bunch of math simpler.  Internal numbers have the following units:

distance: millimeters
time: milliseconds
speed: millimeters/millsecond

Note that internal speed is equivlent to m/s (yay metric!).

The routines for editing leg data still use values in miles, mph as the user needs to edit those values and it doesn't make sense to continually convert
everytime a button is pressed.  Coversion is done entering and exiting the leg edit state.

+++ race_computer/gps.h
@@ -23,8 +23,10 @@ struct gpsDataStruct {
   bool timeValid;
   bool fixValid;
   orcTime_t gpsTime;
+  //GPS speed in um/ms
   double speed;
-  unsigned long distance;
+  //GPS odometer distance in um.
+  double distance;
 };
